refactor: 필드명 수정
