+++ src/test/java/roomescape/reservation/repository/ReservationRepositoryImplTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.database;
+package roomescape.reservation.repository;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.dao.IncorrectResultSizeDataAccessException;
+import roomescape.reservation.database.ReservationRepositoryImpl;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.fixture.ReservationFixture;
 
@@ -17,7 +18,7 @@ import java.util.List;
 class ReservationRepositoryImplTest {
 
     @Autowired
-    private ReservationRepositoryImpl db;
+    private ReservationRepositoryImpl repository;
 
     @DisplayName("존재하지 않는 예약 ID로 조회하면 예외가 발생한다.")
     @Test
@@ -34,12 +35,12 @@ class ReservationRepositoryImplTest {
         List<Reservation> reservations = List.of(reservation1, reservation2);
 
         for (Reservation reservation : reservations) {
-            db.add(reservation);
+            repository.add(reservation);
         }
 
         // when & then
         Assertions.assertThatCode(
-                () -> db.findById(Long.MAX_VALUE)
+                () -> repository.findById(Long.MAX_VALUE)
         ).isInstanceOf(IncorrectResultSizeDataAccessException.class);
     }
 }
